infra/lambda/drift_check/requirements.txt
snowflake-connector-python==3.10.0
boto3>=1.34
